requirements.txt
joblib==1.5.3
xlrd==2.0.1
python-dotenv==1.2.2
psycopg2-binary==2.9.11
streamlit==1.52.2
numpy==2.3.5
pandas==2.3.3
python-dateutil==2.9.0.post0
pytz==2025.2
scikit-learn==1.8.0
scipy==1.16.3
six==1.17.0
threadpoolctl==3.6.0
tzdata==2025.3